+++ sonar-delphi-plugin/src/main/java/au/com/integradev/delphi/DelphiSensor.java
@@ -114,6 +114,7 @@ public class DelphiSensor implements Sensor {
             .sourceFiles(sourceFiles)
             .referencedFiles(referencedFiles)
             .encoding(delphiProjectHelper.encoding())
+            .ansiCharset(delphiProjectHelper.getAnsiCharset())
             .searchPath(searchPath)
             .conditionalDefines(delphiProjectHelper.getConditionalDefines())
             .unitScopeNames(delphiProjectHelper.getUnitScopeNames())
@@ -131,6 +132,7 @@ public class DelphiSensor implements Sensor {
     DelphiFileConfig config =
         DelphiFile.createConfig(
             delphiProjectHelper.encoding(),
+            delphiProjectHelper.getAnsiCharset(),
             preprocessorFactory,
             typeFactory,
             searchPath,